+++ reactionThreads.js
@@ -1,8 +1,9 @@
 module.exports = function ({ bot, config, commands, threads }) {
   const fs = require("fs");
-  const pluginVersion = "1.1";
-  const changelogUrl = "=> https://daark.de/RTCL <="
+  const pluginVersion = "1.2";
+  const changelogUrl = "=> https://daark.de/RTCL <=";
   let reactions = [];
+  const emptyResponse = ["none", "nothing", "empty", "null", "-"];
 
   // Check if ownerId is specified in the config, warn otherwise
   const ownerId = config["reactionThreads-ownerId"];
@@ -14,7 +15,7 @@ module.exports = function ({ bot, config, commands, threads }) {
   // Load the suffix for the json file, if one exists (used for multiple bot instances running from the same folder)
   const jsonSuffix = config["reactionThreads-suffix"] ? config["reactionThreads-suffix"] : "";
 
-  // Warn the user not to delete the file in case it doesnt exist (basically a first-use check)
+  // Warn the user not to delete the file in case it doesn't exist (basically a first-use check)
   if (!fs.existsSync(`./ReactionThreadsData${jsonSuffix}.json`)) {
     console.info(
       `[ReactionThreads] A ReactionThreadsData${jsonSuffix}.json file will be created when using this plugin. Please do not modify or delete this file or reactions you set up will cease to function.`,
@@ -49,10 +50,29 @@ module.exports = function ({ bot, config, commands, threads }) {
     return null;
   };
 
-  const isOwner = function (message) {
+  /**
+   * Checks whether or not passed parameters has response
+   * @param {string} channelId The ID of the channel for which to check
+   * @param {string} messageId The ID of the message for which to check
+   * @param {string} emoji The stringified emoji for which to check (i.e. <:test:108552944961454080>)
+   * @returns full reaction if valid, null if not
+   */
+  const hasResponse = function (channelId, messageId, emoji) {
+    const reaction = isValidReaction(channelId, messageId, emoji);
+    if (reaction && reaction.response) {
+      return reaction;
+    }
+    return null;
+  };
+
+  /**
+   * Checks whether or not the user invoking a command is authorized or not
+   * @param {*} msg the message to check permissions for
+   */
+  const isOwner = function (msg) {
     if (typeof ownerId === "undefined") return true;
-    return message.member.id === ownerId ? true : message.member.roles.includes(ownerId);
-  }
+    return msg.member.id === ownerId ? true : msg.member.roles.includes(ownerId);
+  };
 
   /**
    * Registers a new reaction for use
@@ -61,7 +81,6 @@ module.exports = function ({ bot, config, commands, threads }) {
    */
   const addReactionCmd = async (msg, args) => {
     if (!isOwner(msg)) return;
-    // Didnt work without JSON.stringify, but if its stupid but it works...
     if (isValidReaction(args.channelId, args.messageId, args.emoji)) {
       msg.channel.createMessage(`⚠️ Unable to add reaction: That reaction already exists on that message!`);
       return;
@@ -80,19 +99,18 @@ module.exports = function ({ bot, config, commands, threads }) {
 
     reactions.push({ ...args });
     saveReactions();
-    msg.channel.createMessage("Succesfully added reaction to message and registered it internally.");
+    msg.channel.createMessage("Successfully added reaction to message and registered it internally.");
   };
 
   /**
-   * Unregisters an existing reaction
+   * Deregister an existing reaction
    * @param {Message} msg The message invoking the command
    * @param {*} args The arguments passed (check registering at bottom)
    */
   const removeReactionCmd = async (msg, args) => {
     if (!isOwner(msg)) return;
-    // Didnt work without JSON.stringify, but if its stupid but it works...
     if (isValidReaction(args.channelId, args.messageId, args.emoji) == null) {
-      msg.channel.createMessage(`⚠️ Unable to remove reaction: That reaction doesnt exist on that message!`);
+      msg.channel.createMessage(`⚠️ Unable to remove reaction: That reaction doesn't exist on that message!`);
       return;
     }
 
@@ -105,14 +123,68 @@ module.exports = function ({ bot, config, commands, threads }) {
 
     reactions.splice(reactions.indexOf({ ...args }), 1);
     saveReactions();
-    msg.channel.createMessage("Succesfully removed reaction from message and de-registered it internally.");
+    msg.channel.createMessage("Successfully removed reaction from message and de-registered it internally.");
+  };
+
+  /**
+   * Registers or updates a reaction response
+   * @param {Message} msg The message invoking the command
+   * @param {*} args The arguments passed (check registering at bottom)
+   */
+  const ReactionRespCmd = async (msg, args) => {
+    if (!isOwner(msg)) return;
+    reaction = isValidReaction(args.channelId, args.messageId, args.emoji);
+    if (reaction == null) {
+      msg.channel.createMessage(`⚠️ Unable to add reaction response: That reaction doesn't exist on that message!`);
+      return;
+    }
+
+    if (args.response) {
+      const response = emptyResponse.includes(args.response.trim()) ? null : args.response.trim();
+      if (response.length > 1500) {
+        msg.channel.createMessage("⚠️ That custom response is too long!");
+        return;
+      }
+
+      reaction.response = response;
+      saveReactions();
+      msg.channel.createMessage("Successfully created/updated reaction response and registered it internally.");
+    } else {
+      msg.channel.createMessage(
+        `The current response for that reaction is: \`${reaction.response ? reaction.response : "None"}\``,
+      );
+    }
+  };
+
+  /**
+   * Lists all (or specific) reactions
+   * @param {*} msg The message which invoked the command
+   * @param {*} args The arguments passed to us, i.e. an ID
+   */
+  const listReactionsCmd = async (msg, args) => {
+    if (!isOwner(msg)) return;
+    const checkId = args.anyId ? args.anyId.trim() : null;
+    let toPost = "Emoji - Channel ID - Message ID - Category ID - First words of response";
+    for (const react of reactions) {
+      if (react.channelId === "version") continue;
+      if (args.anyId && !(react.categoryId === checkId || react.channelId === checkId || react.messageId === checkId))
+        continue;
+
+      const reactionText = `\n${react.emoji} - ${react.channelId} - ${react.messageId} - ${
+        react.categoryId ? react.categoryId : "None"
+      } - ${react.response ? react.response.substring(0, 25) + "..." : "None"}`;
+
+      toPost += reactionText;
+    }
+
+    msg.channel.createMessage(toPost);
   };
 
   /**
    * Handles any reaction added within a guild. If it is a registered reaction, create a thread if none exists
    * @param {Message} message The message that got reacted to
    * @param {Emoji} emoji The emoji used to react
-   * @param {Member} reactor The memeber object of the person reacting
+   * @param {Member} reactor The member object of the person reacting
    */
   const onReactionAdd = async (message, emoji, reactor) => {
     if (reactor.user.bot || !reactor.guild) return;
@@ -130,13 +202,22 @@ module.exports = function ({ bot, config, commands, threads }) {
         ? config.responseMessage.join("\n")
         : config.responseMessage;
       const postToThreadChannel = config.showResponseMessageInThreadChannel;
+
       await newThread.postSystemMessage(
         `:gear: **ReactionThreads:** Thread opened because of reaction ${stringifiedEmoji} to https://discord.com/channels/${
           message.channel.guild.id
         }/${message.channel.id}/${message.id}${toPing != null ? " <@&" + toPing + ">" : ""}`,
         { allowedMentions: { roles: [toPing] } },
       );
-      newThread.sendSystemMessageToUser(responseMessage, { postToThreadChannel });
+
+      newThread
+        .sendSystemMessageToUser(reaction.response ? reaction.response : responseMessage, { postToThreadChannel })
+        .catch((e) => {
+          // Ideally this will be fixed upstream at some point
+          newThread.postSystemMessage(
+            "⚠️ **ReactionThreads:** Could not open DMs with the user. They may have blocked the bot or set their privacy settings higher.",
+          );
+        });
 
       try {
         await bot.removeMessageReaction(message.channel.id, message.id, stringifiedEmoji.replace(">", ""), reactor.id);
@@ -153,7 +234,6 @@ module.exports = function ({ bot, config, commands, threads }) {
     }
   };
 
-
   //#region versioncheck
   // Check the plugin version and notify of any updates that happened
   let foundVersion = null;
@@ -165,7 +245,9 @@ module.exports = function ({ bot, config, commands, threads }) {
   }
 
   if (foundVersion != null && foundVersion != pluginVersion) {
-    console.info(`[ReactionThreads] Plugin updated to version ${pluginVersion}, please read the changelog at ${changelogUrl} as there may be important or breaking changes!`);
+    console.info(
+      `[ReactionThreads] Plugin updated to version ${pluginVersion}, please read the changelog at ${changelogUrl} as there may be important or breaking changes!`,
+    );
     reactions.splice(reactions.indexOf({ channelId: "version", messageId: foundVersion }), 1);
     reactions.push({ channelId: "version", messageId: pluginVersion });
     saveReactions();
@@ -173,6 +255,7 @@ module.exports = function ({ bot, config, commands, threads }) {
     reactions.push({ channelId: "version", messageId: pluginVersion });
     saveReactions();
   }
+  //#endregion versioncheck
 
   //#region registering
   // Register all commands and listeners
@@ -200,6 +283,17 @@ module.exports = function ({ bot, config, commands, threads }) {
   );
 
   commands.addInboxServerCommand(
+    "rtResponse",
+    [
+      { name: "channelId", type: "string", required: true },
+      { name: "messageId", type: "string", required: true },
+      { name: "emoji", type: "string", required: true },
+      { name: "response", type: "string", required: false, catchAll: true },
+    ],
+    ReactionRespCmd,
+  );
+
+  commands.addInboxServerCommand(
     "rtRemove",
     [
       { name: "channelId", type: "string", required: true },
@@ -208,6 +302,8 @@ module.exports = function ({ bot, config, commands, threads }) {
     ],
     removeReactionCmd,
   );
+
+  commands.addInboxServerCommand("rtList", [{ name: "anyId", type: "string", required: false }], listReactionsCmd);
 
   bot.on("messageReactionAdd", onReactionAdd);
   //#endregion